+++ test/test_helper.rb
@@ -4,4 +4,4 @@ $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'phlex/stimulus'
 require 'minitest/autorun'
 require 'shoulda-context'
-require 'byebug'
+require 'debug'